// src/Momentum.Frontend/apps/code-innovation-backlog/src/generated/models/Cycai_momentumsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cycai_momentumscycai_targettype = {
  100000000: 'Idea',
  100000001: 'Solution'
} as const;
export type Cycai_momentumscycai_targettype = keyof typeof Cycai_momentumscycai_targettype;
export const Cycai_momentumsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cycai_momentumsstatecode = keyof typeof Cycai_momentumsstatecode;
export const Cycai_momentumsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cycai_momentumsstatuscode = keyof typeof Cycai_momentumsstatuscode;

export interface Cycai_momentumsBase {
  cycai_activeusecount?: number;
  cycai_adoptioncount?: number;
  cycai_adoptions30d?: number;
  cycai_calculatedon?: string;
  cycai_commentcount?: number;
  cycai_completedusecount?: number;
  cycai_contributorcount?: number;
  cycai_demandrank?: number;
  cycai_linkedcount?: number;
  cycai_momentumid: string;
  cycai_momentumscore?: number;
  cycai_name: string;
  cycai_previousdemandrank?: number;
  cycai_targetid: number;
  cycai_targetkey: string;
  cycai_targettype: Cycai_momentumscycai_targettype;
  cycai_teamcount?: number;
  cycai_votecount?: number;
  cycai_votes30d?: number;
  cycai_votes7d?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cycai_momentumsstatecode;
  statuscode?: Cycai_momentumsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cycai_momentums extends Cycai_momentumsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cycai_targettypename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}